+++ pinot-core/src/main/java/org/apache/pinot/core/data/table/SimpleIndexedTable.java
@@ -133,15 +133,6 @@ public class SimpleIndexedTable extends IndexedTable {
   }
 
   @Override
-  public boolean merge(Table table) {
-    Iterator<Record> iterator = table.iterator();
-    while (iterator.hasNext()) {
-      upsert(iterator.next());
-    }
-    return true;
-  }
-
-  @Override
   public int size() {
     return _lookupMap.size();
   }
@@ -170,10 +161,5 @@ public class SimpleIndexedTable extends IndexedTable {
     if (_iterator == null) {
       _iterator = _lookupMap.values().iterator();
     }
-  }
-
-  @Override
-  public DataSchema getDataSchema() {
-    return _dataSchema;
   }
 }
